backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
websockets==12.0
pytest==7.4.3
httpx==0.25.1
redis==5.0.1
numpy==1.26.1
pandas==2.1.2 